FIX Unset parent mime part body, not just children
The message itself is a MimePart, where the PGP signer will set
the contents as children (and clear existing ones), it wrongly did
not unset the body - meaning that a message as a single body was to
be transmitted as plain text as the message body, then also encrypted
as an attachemnt. This made PGP encryption pointeless, and is a rather
glaring bug.